luci-app-passwall: optimize and improve
